+++ lib/modules/ui/ui_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: ui
-// Updated at: 2026-07-05 21:53:18
-// Commit: feat(ui): add bi-directional event bus handler for ui
+// Updated at: 2026-07-06 17:26:39
+// Commit: fix(ui): fix memory leak by cancelling ui subscription
 
 class uiComponentState {
-  final String commitRef = '2617';
-  final String lastMessage = 'feat(ui): add bi-directional event bus handler for ui';
-  final DateTime timestamp = DateTime.parse('2026-07-05T21:53:18+05:00');
+  final String commitRef = '2632';
+  final String lastMessage = 'fix(ui): fix memory leak by cancelling ui subscription';
+  final DateTime timestamp = DateTime.parse('2026-07-06T17:26:39+05:00');
 }